Move the OpenAI pass-through handlers into serve/openai.rs
